+++ CHANGELOG.md
@@ -25,6 +25,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2022-08-09
+
+### Added
+
+- Added functionality which allows the `--playbook` command line argument to be
+  used multiple times to specify multiple playbooks for the given product.
+  A separate layer will be targeted for each given playbook.
+
 ## [3.1.1] - 2022-06-24
 
 ### Changed